demos/07_containerized_workers/apps/task_manager/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
kombu==5.2.4
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.20
task-shared @ file:///home/student/stock-data-tool/apps/task_manager/task_shared-0.1.0-py3-none-any.whl#sha256=fd8085fb4c5ac5b279a3cc2c1051daf88a50fa70d8f203dac1b3fec29ba039a8
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
